Archives a file
This script will archive a file in a subdirectory "old" in the current directory.  It will add a date stamp to it.  If subdirectory "old" doesn't exist, it will be created, and then the file will be put there.
For example:  arc myFile.ext will be saved as old\myFile_yyyymmdd.ext (where yyyymmdd is the current date).